Changes to bean AI
I've implemented that the beans constantly keep the hiding spot between them and the player, the speed that they do so can be adjusted in the inspector for difficulty adjustment. Also improved the switching hiding spot logic, the beans now have a more randomised setting for when each bean moves, the number that can be moved at one time can be adjusted in the inspector also on the AIManager game object in the hierarchy.