Added new beans and their movement  animation
Added the new bean prefab with it's movement animation. Works so that the beans only bounce when on the move and stop bouncing when in hiding spots. Fixed bug of not being able to pickup bean after dropping(Issue was in InteractableObjects cs. Implemented stop for sound when bean is picked up, NPC_AI cs was being set to inactive in the InteractableObjects cs which was preventing the code from updating when the bean was picked up.