Fixes to ai implementation
Fixed trap spawn. Traps now spawn on a fixed location, preventing the issue of spawning off mesh.

Fixed issue where traps are picked up when a bean is inside so only the bean is picked up.

Fixed ! upgrade logic so it only activates when the bean is within a defined range of the player and does not appear constantly.

Fixed issue where all beans start moving around at once, issue was that there wasn't a check if the beans were running, now beans only switch hiding spots if there are non that are currently running.