Bug fix
Removed collision from teacups above the coffee machine due to the collider randomly distributing across the shop.

Resized colliders encapsulating the hopper to reduce the amount of shattered bean bits that can escape while blending.

Resized the collider for the button and customer window to reduce the risk of missing it